backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-ninja==0.10.2
pydantic==1.7.3
pytz==2021.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
